Securing the views for product admin

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 
@@ -74,24 +75,15 @@ def product_detail(request, product_id):
 
     return render(request, 'products/product_detail.html', context)
 
-# def CreateProductView(request):
-#     form = CreateProduct()
 
-#     if request.method == "POST":
-#         form = CreateProduct(request.POST,request.FILES)
-#         if forms.is_valid():
-#             upload_form = form.save(commit = False)
-#             upload_form.created_by = request.user
-#             upload_form.save()
-
-#     context = {
-#         "form":form
-#     }
-#     return render(request,"products/upload_form.html",context)
-
-
+@login_required
 def add_product(request):
     """ Add a product to the store """
+    # Need to change this to include is_vendor
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -111,8 +103,14 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    # Need to change this to include is_vendor
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -135,8 +133,14 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ Delete a product from the store """
+    # Need to change this to include is_vendor
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that.')
+        return redirect(reverse('home'))
+        
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
